refactor: onSubmit 함수들 RacingGame class로 이동
