+++ src/index.js
@@ -1,5 +1,5 @@
 import RacingGame from './RacingGame.js';
-import initForms from './utils/forms.js';
+import initForms from './utils/initForms.js';
 
 const racingGame = new RacingGame();
